adjust types, rename helper fns
